+++ Models/Location.cs
@@ -5,7 +5,7 @@ namespace CargoCult.Models
 {
     public class Location
     {
-        public int LocationID { get; set; }
+        public long LocationID { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public double Latitude { get; set; }
